[FEAT/#7] 회원가입 레트로핏 코드 수정
