+++ app/build.gradle
@@ -80,7 +80,7 @@ dependencies {
     // define a BOM and its version
     implementation(platform("com.squareup.okhttp3:okhttp-bom:4.10.0"))
 
-    // define any required OkHttp artifacts without version (위의 BOM에 맞게 버전 적용)
+    // define any required OkHttp artifacts without version
     implementation("com.squareup.okhttp3:okhttp")
     implementation("com.squareup.okhttp3:logging-interceptor")
