+++ app/src/main/java/org/android/go/sopt/data/remote/model/Request.kt
@@ -1,0 +1,16 @@
+package org.android.go.sopt.data.remote.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestSignUpDto(
+    @SerialName("id")
+    val id: String,
+    @SerialName("password")
+    val password: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("skill")
+    val skill: String,
+)
