+++ app/src/main/java/org/android/go/sopt/data/remote/model/Response.kt
@@ -1,19 +1,7 @@
 package org.android.go.sopt.data.remote.model
 
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
-
-@Serializable
-data class RequestSignUpDto(
-    @SerialName("id")
-    val id: String,
-    @SerialName("password")
-    val password: String,
-    @SerialName("name")
-    val name: String,
-    @SerialName("skill")
-    val skill: String,
-)
+import kotlinx.serialization.Serializable
 
 @Serializable
 data class ResponseSignUpDto(
